requirements.txt
requests==2.27.1
flask==2.3.2
python-dotenv==0.19.2